fix: fixed crud endpoints from authors

+++ api/authors/controller.py
@@ -1,28 +1,24 @@
-from api.config.app import app, Response
-from api.authors import services
-
-@app.route(path="/authors", methods=['GET'])
-def get_all(request):
-    authors = services.get_authors()
-    return Response(json={'authors':authors}, status=200)
+from config.app import app
+from authors import services
 
 
-@app.route(path="/authors", methods=['POST'])
-def create(request):
-    new_author = services.create_author(request.json)
-    return Response(json={"author":new_author}, status=201)
+@app.route(path="/authors", methods=['GET', 'POST'])
+def books(request):
+    if request.method == 'GET':
+        authors = services.get_authors()
+        return {'authors':authors}, 200
+    if request.method == 'POST':
+        new_author = services.create_author(request.json)
+        return {"author":new_author}, 201
 
-@app.route(path="/authors/{id}", methods=['GET'])
-def get_one(request, id):
-    author = services.get_author(id)
-    return Response(json={'author':author}, status=200)
-
-@app.route(path="/authors/{id}", methods=['PUT'])
-def update(request, id):
-    author = services.update_author(id)
-    return Response(json={'author':author}, status=200)
-
-@app.route(path="/authors/{id}", methods=['DELETE'])
-def delete(request, id):
-    result = services.delete_author(id)
-    return Response(json=result, status=200)
+@app.route(path="/authors/{id}", methods=['GET', 'PUT','DELETE'])
+def book(request, id):
+    if request.method == 'GET':
+        author = services.get_author(id)
+        return {'author':author}, 200
+    if request.method == 'PUT':
+        author = services.update_author(id, request.json)
+        return {'author':author},200
+    if request.method == 'DELETE':
+        result = services.delete_author(id)
+        return result, 200
